internal/terminalgame: refactoring
