--- internal/terminalgame/game/doc.go
@@ -1,2 +1,0 @@
-// Package game contains terminal game implementation
-package game
