+++ internal/terminalgame/gameimpl/doc.go
@@ -1,0 +1,2 @@
+// Package gameimpl contains terminal game implementation
+package gameimpl
