+++ internal/terminalgame/gameimpl/game.go
@@ -1,11 +1,12 @@
-package game
+package gameimpl
 
 import (
 	"bufio"
 	"fmt"
-	"github.com/gucio321/tic-tac-go/pkg/core/board/letter"
 	"log"
 	"os"
+
+	"github.com/gucio321/tic-tac-go/pkg/core/board/letter"
 
 	"github.com/gucio321/go-clear"
 	"github.com/gucio321/terminalmenu/pkg/menuutils"
