+++ internal/terminalgame/gameimpl/user_api.go
@@ -1,4 +1,4 @@
-package game
+package gameimpl
 
 import (
 	"errors"
