+++ internal/terminalgame/menu/menu.go
@@ -3,6 +3,7 @@ package menu
 import (
 	"errors"
 	"fmt"
+	"github.com/gucio321/go-clear"
 	"log"
 	"math/rand"
 	"runtime"
@@ -16,13 +17,11 @@ import (
 
 	"github.com/gravestench/osinfo"
 
-	"github.com/gucio321/go-clear"
 	terminalmenu "github.com/gucio321/terminalmenu/pkg"
 	"github.com/gucio321/terminalmenu/pkg/menuutils"
 
-	gameimpl "github.com/gucio321/tic-tac-go/internal/terminalgame/game"
+	"github.com/gucio321/tic-tac-go/internal/terminalgame/gameimpl"
 	"github.com/gucio321/tic-tac-go/pkg/core/board"
-	"github.com/gucio321/tic-tac-go/pkg/core/players/player"
 	"github.com/gucio321/tic-tac-go/pkg/game"
 )
 
@@ -84,7 +83,7 @@ func (m *Menu) Run() {
 }
 
 func (m *Menu) runPVP() {
-	pvp := game.NewTTG(m.width, m.height, m.chainLen, game.PlayerTypeHuman, player.PlayerPerson)
+	pvp := gameimpl.NewTTG(m.width, m.height, m.chainLen, game.PlayerTypeHuman, game.PlayerTypeHuman)
 	pvp.Run()
 }
 
@@ -96,16 +95,16 @@ func (m *Menu) runPVC() {
 
 	switch r {
 	case 0:
-		g = game.NewTTG(m.width, m.height, m.chainLen, player.PlayerPerson, player.PlayerPC)
+		g = gameimpl.NewTTG(m.width, m.height, m.chainLen, game.PlayerTypeHuman, game.PlayerTypePC)
 	case 1:
-		g = game.NewTTG(m.width, m.height, m.chainLen, player.PlayerPC, player.PlayerPerson)
+		g = gameimpl.NewTTG(m.width, m.height, m.chainLen, game.PlayerTypePC, game.PlayerTypeHuman)
 	}
 
 	g.Run()
 }
 
 func (m *Menu) runDemo() {
-	demo := game.NewTTG(m.width, m.height, m.chainLen, player.PlayerPC, player.PlayerPC)
+	demo := gameimpl.NewTTG(m.width, m.height, m.chainLen, game.PlayerTypePC, game.PlayerTypePC)
 	demo.Run()
 }
